+++ R/compute-sequence.R
@@ -4,7 +4,7 @@
     chroms <- c()
     files <- c()
     tmp.dirname <- tempfile(pattern = "", tmpdir = paste(groot, "/downloads", sep = ""))
-    if (!dir.create(tmp.dirname, recursive = TRUE, mode = "0777")) {
+    if (!.gwith_umask(dir.create(tmp.dirname, recursive = TRUE, mode = "0777"))) {
         stop(sprintf("Failed to create a directory %s", tmp.dirname), call. = FALSE)
     }
 
@@ -94,7 +94,7 @@
         {
             if (grepl("\\.gz$", fasta, perl = TRUE)) {
                 tmp.dirname <- tempfile(pattern = "", tmpdir = paste(groot, "/downloads", sep = ""))
-                if (!dir.create(tmp.dirname, recursive = TRUE, mode = "0777")) {
+                if (!.gwith_umask(dir.create(tmp.dirname, recursive = TRUE, mode = "0777"))) {
                     stop(sprintf("Failed to create temp directory %s", tmp.dirname), call. = FALSE)
                 }